changes to the independent sampler (#505)
* we now set radii using BSE and we now sample prob first and use that to limit the eccentricity sampling for the independent sampler
